back/requirements.txt
openai
python-dotenv
firebase-functions